PLATUI-1289: Refactored Rich implicits to separate package and refactored classes duplicated from play-frontend-govuk
