+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.85.0] - 2021-07-27
+## [0.87.0] - 2021-07-28
+
+### Changed
+
+- Updated version of `play-frontend-govuk` to 0.82.0
+- Refactored Rich implicits
+- Removed duplicated code RichHtml, RichString, and RichOptionString
+
+### Compatible with
+
+- [hmrc/hmrc-frontend v2.0.0](https://github.com/hmrc/hmrc-frontend/releases/tag/v2.0.0)
+- [hmrc/play-frontend-govuk v0.82.0](https://github.com/hmrc/play-frontend-govuk/releases/tag/v0.82.0)
+- [alphagov/govuk-frontend v3.13.0](https://github.com/alphagov/govuk-frontend/releases/tag/v3.13.0)
+
+## [0.86.0] - 2021-07-27
 
 ### Changed
 
@@ -16,6 +30,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - [hmrc/hmrc-frontend v2.0.0](https://github.com/hmrc/hmrc-frontend/releases/tag/v2.0.0)
 - [hmrc/play-frontend-govuk v0.81.0](https://github.com/hmrc/play-frontend-govuk/releases/tag/v0.81.0)
+- [alphagov/govuk-frontend v3.13.0](https://github.com/alphagov/govuk-frontend/releases/tag/v3.13.0)
+
+## [0.85.0] - 2021-07-23
+
+### Changed
+
+- Added missing space before a nonce attribute in `hmrcHead` and `hmrcScripts`
+
+### Compatible with
+
+- [hmrc/hmrc-frontend v2.0.0](https://github.com/hmrc/hmrc-frontend/releases/tag/v2.0.0)
+- [hmrc/play-frontend-govuk v0.80.0](https://github.com/hmrc/play-frontend-govuk/releases/tag/v0.80.0)
 - [alphagov/govuk-frontend v3.13.0](https://github.com/alphagov/govuk-frontend/releases/tag/v3.13.0)
 
 ## [0.84.0] - 2021-07-21
